requirements.txt
streamlit
openai
moviepy
gtts
python-dotenv
Pillow
requests